ai-fundamentals-series/shared/requirements.txt
# Dependencies for Blog 2: Prompt Engineering Examples
#
# Core LLM provider libraries
openai>=1.0.0                    # OpenAI GPT models
google-cloud-aiplatform>=1.38.0  # Google VertexAI
anthropic>=0.7.0                 # Anthropic Claude models
ollama>=0.1.0                    # Ollama local models

# Environment and configuration
python-dotenv>=1.0.0             # Load environment variables from .env file
pydantic>=2.0.0                  # Data validation and settings management

# Utilities
requests>=2.31.0                 # HTTP requests for custom API calls
tenacity>=8.2.0                  # Retry mechanisms for API calls
tiktoken>=0.5.0                  # Token counting for OpenAI models
colorama>=0.4.6                  # Colored terminal output for demos

# Development and testing
pytest>=7.4.0                   # Testing framework
black>=23.0.0                   # Code formatting
isort>=5.12.0                   # Import sorting